Implemented saving homeworks into database - Reworked homework database handler - Changed id system (local database and from internet database) - Filter now has better performance in homeworkview

+++ app/src/main/java/de/gymnasium_lappersdorf/gymlapapp/HausaufgabenPlaner/JsonHandler.java
@@ -50,8 +50,8 @@ public class JsonHandler {
     }
 
 
-    public int getID(int position) {
-        return getInt(position, "id");
+    public long getID(int position) {
+        return getLong(position, "id");
     }
 
     public String getFach(int position) {
